fix: refactor api package
