+++ src/main/java/ru/mirea/megatracker/api/coin/CoinPriceData.java
@@ -1,4 +1,5 @@
-package ru.mirea.megatracker.api;
+package ru.mirea.megatracker.api.coin;
+
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -7,17 +8,13 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-
-@Getter
 @Setter
+@Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class Coin {
+public class CoinPriceData {
 
-    @JsonProperty(value = "RAW")
-    private CoinPriceData coinPriceData;
-
-    @JsonProperty(value = "CoinInfo")
-    private CoinInfo coinInfo;
+    @JsonProperty(value = "USD")
+    private PriceInfoUSD priceInfoUSD;
 }
