+++ src/main/java/ru/mirea/megatracker/services/CoinService.java
@@ -4,10 +4,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.reactive.function.client.WebClient;
-import ru.mirea.megatracker.api.Coin;
-import ru.mirea.megatracker.api.CoinHistoryPrice;
-import ru.mirea.megatracker.api.CoinInfo;
-import ru.mirea.megatracker.api.CoinPriceData;
+import ru.mirea.megatracker.api.coin.Coin;
+import ru.mirea.megatracker.api.coin.CoinHistoryPrice;
+import ru.mirea.megatracker.api.coin.CoinInfo;
+import ru.mirea.megatracker.api.coin.CoinPriceData;
 import ru.mirea.megatracker.api.response.ExchangePairApiResponse;
 import ru.mirea.megatracker.api.response.HistoryApiResponse;
 import ru.mirea.megatracker.api.response.TopListApiResponse;
